Inventory complete

+++ ReStore/API/Controllers/ProductsController.cs
@@ -7,6 +7,7 @@ using API.DTOs;
 using API.Entities;
 using API.Extensions;
 using API.RequestHelpers;
+using API.Services;
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -18,8 +19,10 @@ namespace API.Controllers
     {
         private readonly StoreContext _context;
         private readonly IMapper _mapper;
-        public ProductsController(StoreContext context, IMapper mapper)
+        private readonly ImageService _imageService;
+        public ProductsController(StoreContext context, IMapper mapper, ImageService imageService)
         {
+            _imageService = imageService;
             _mapper = mapper;
             _context = context;
         }
@@ -62,9 +65,17 @@ namespace API.Controllers
 
         [Authorize(Roles = "Admin")]
         [HttpPost]
-        public async Task<ActionResult<Product>> CreateProduct(CreateProductDto productDto)
+        public async Task<ActionResult<Product>> CreateProduct([FromForm] CreateProductDto productDto)
         {
             var product = _mapper.Map<Product>(productDto);
+            if (productDto.File != null)
+            {
+                var imageResult = await _imageService.AddImageAsync(productDto.File);
+                if (imageResult.Error != null) return BadRequest(new ProblemDetails { Title = imageResult.Error.Message });
+
+                product.PictureUrl = imageResult.SecureUrl.ToString();
+                product.PublicId = imageResult.PublicId;
+            }
             _context.Products.Add(product);
 
             var result = await _context.SaveChangesAsync() > 0;
@@ -76,7 +87,7 @@ namespace API.Controllers
 
         [Authorize(Roles = "Admin")]
         [HttpPut]
-        public async Task<ActionResult> UpdateProduct(UpdateProductDto productDto)
+        public async Task<ActionResult<Product>> UpdateProduct([FromForm] UpdateProductDto productDto)
         {
             var product = await _context.Products.FindAsync(productDto.Id);
 
@@ -84,9 +95,20 @@ namespace API.Controllers
 
             _mapper.Map(productDto, product);
 
+            if (productDto.File != null)
+            {
+                var imageResult = await _imageService.AddImageAsync(productDto.File);
+                if (imageResult.Error != null) return BadRequest(new ProblemDetails { Title = imageResult.Error.Message });
+
+                if (!string.IsNullOrEmpty(product.PublicId)) await _imageService.DeleteImageAsync(product.PublicId);
+
+                product.PictureUrl = imageResult.SecureUrl.ToString();
+                product.PublicId = imageResult.PublicId;
+            }
+
             var result = await _context.SaveChangesAsync() > 0;
 
-            if (result) return NoContent();
+            if (result) return Ok(product);
 
             return BadRequest(new ProblemDetails { Title = "Problem updating product" });
         }
@@ -96,6 +118,9 @@ namespace API.Controllers
         public async Task<ActionResult> DeleteProduct(int id)
         {
             var product = await _context.Products.FindAsync(id);
+            if (product == null) return NotFound();
+            if (!string.IsNullOrEmpty(product.PublicId)) await _imageService.DeleteImageAsync(product.PublicId);
+
             _context.Products.Remove(product);
             var result = await _context.SaveChangesAsync() > 0;
